RBAC for HR * Senior Management
Added role based access control for HR and senior management with references to the position and role of the session. Since HR and senior management needs to see the list and number of staff in office & home, this will only pull approved requests.

Note: Core functionality reqs states they are looking for number/list. Can write another SQL code to pull distinct counts of WFH requests to tally the amount of people in office by dept. For now, only shows all WFH requests in the org.