New view own team schedule for all users. List of people in office for manager view.
Added a new webpage called viewownteamschedule and added a button to redirect to it for all roles. This view shows the presence of the whole team in office (by team it is people who report to the same manager). Manager view, added a list which tallies who is in office by setting a default everyone in office, iterating through each WFH request and pulling out said person before returning all that's left.